Hamburger menu dropdown items filled and logged in and signed up nav bar components split into seperate files for readability
